Start on a simple `npm` function.
The motivation is that I work on various projects and some use `npm` and some use `yarn` and it is very annoying to remember which use which and I end up using the wrong one and accidentally creating lock files and such.

The purpose of this function is to let me use the npm CLI syntax for my most common things and have Fish translate them to yarn commands automatically if I'm in a yarn project.